# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Risk-Aware Lending Protocol API
  version: 1.0.0
  description: |
    Read API over the lending protocol. Two sources, never mixed:

    * **Live state** (markets, accounts, health factors, oracle health) is read from the chain
      through `PoolLens` on every request (2 s cache for market-wide reads).
    * **History** (events, liquidations, rate history, analytics) comes from PostgreSQL, filled
      by the reorg-aware indexer. Risk classifications and alerts come from the risk monitor.

    Every on-chain quantity is returned as an exact integer string (`raw`) plus display numbers.
    Real-time updates: connect to `GET /ws` (WebSocket) for committed index batches, reorgs and
    risk-alert changes.
servers:
  - url: http://localhost:4400
...
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: INVALID_ADDRESS }
            message: { type: string }
...
    RiskLevel:
      type: string
      enum: [SAFE, WARNING, DANGER, LIQUIDATABLE, UNKNOWN]
      description: "SAFE HF>=1.5 or no debt; WARNING 1.2-1.5; DANGER 1.0-1.2; LIQUIDATABLE <1; UNKNOWN = an exposed oracle is down"
...
paths:
...
  /risk/accounts:
    get:
      tags: [risk]
      summary: Indebted accounts from the latest monitor sweep, riskiest first
      parameters:
        - { name: level, in: query, schema: { $ref: "#/components/schemas/RiskLevel" } }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200": { description: Accounts }
        "400": { $ref: "#/components/responses/BadRequest" }
  /risk/alerts:
    get:
      tags: [risk]
      summary: Monitor alerts
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, resolved, all], default: open } }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200": { description: Alerts }
        "400": { $ref: "#/components/responses/BadRequest" }
...